Add files via upload
This project comprises of SQL codes and visualizations and web scraping done with Power Bi. 
It is a 4-Month Amazon Revenue Data that shows key performance index and indices. 
With Microsoft Power Bi, raw data, cleaned and queried in SQL Server is now well visualized to make more sense for easy assessment and comprehension